Allow legacy custom scheme registration on non-main threads https://bugs.webkit.org/show_bug.cgi?id=217656 <rdar://problem/70228036>
Patch by Alex Christensen <achristensen@webkit.org> on 2020-10-13
Reviewed by Brady Eidson.

r267802 was quite necessary, but it exposed an internal client calling our API on a background thread,
which crashes now.  Make it not crash.  There is a race condition in their registration, but there already was.

* UIProcess/API/Cocoa/WKBrowsingContextController.mm:
(+[WKBrowsingContextController registerSchemeForCustomProtocol:]):
(+[WKBrowsingContextController unregisterSchemeForCustomProtocol:]):

Identifier: 230410@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268401 268f45cc-cd09-0410-ab3c-d52691b4dbfc