It should be possible to create an AudioWorkletNode for a context that has finished rendering https://bugs.webkit.org/show_bug.cgi?id=217666
Reviewed by Eric Carlson.

LayoutTests/imported/w3c:

Rebaseline WPT test that is now passing.

* web-platform-tests/webaudio/the-audio-api/the-audioworklet-interface/suspended-context-messageport.https-expected.txt:

Source/WebCore:

It should be possible to create an AudioWorkletNode for a context that has finished rendering.
We were previously throwing an exception when trying to create an AudioWorkletNode for a context
whose state is "closed" (OfflineAudioContext's state becomes closed as soon as rendering
finishes). This was causing us to fail a WPT test and did not match Blink's behavior.

No new tests, rebaselined existing test.

* Modules/webaudio/AudioWorkletNode.cpp:
(WebCore::AudioWorkletNode::create):


Identifier: 230421@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268424 268f45cc-cd09-0410-ab3c-d52691b4dbfc