[WebIDL] Bindings do not support two attributes or functions with the same name only differing by case of first character https://bugs.webkit.org/show_bug.cgi?id=217776
Reviewed by Chris Dumez.

Due to needing to support cases like CSSStyleDeclaration#webkitAlignSelf and CSSStyleDeclaration#WebkitAlignSelf,
the naming convention for the c-wrapper functions in the generated bindings had to be updated to not ignore the
case of the leading character. The new convention uses a leading _ (unless the identifier already starts with one)
changing a function such as jsDOMWindowInstanceFunctionShowModalDialog to jsDOMWindowInstanceFunction_showModalDialog.

* bindings/js/JSDOMWindowCustom.cpp:
* bindings/js/JSLocationCustom.cpp:
Update for new c-function naming convention.

* bindings/scripts/CodeGeneratorJS.pm:
(MangleAttributeOrFunctionName):
(GetAttributeGetterName):
(GetAttributeSetterName):
(GetFunctionName):
Add a leading underscore to all mangled attribute and function names and remove
leading capitalization. This allows us to have attributes and functions with the
same name except for the capitalization of the first letter.

* bindings/scripts/test/JS/*:
Update bindings test results for new c-function naming scheme.


Identifier: 230514@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268551 268f45cc-cd09-0410-ab3c-d52691b4dbfc