Ensure the tests are run with US/Pacific time zone https://bugs.webkit.org/show_bug.cgi?id=186612
Reviewed by Michael Catanzaro.

Some tests fail if the time zone is not set to US/Pacific, and
this causes issues for contributors living outside of that timezone.
Ideally we should fix those tests, but in the meantime setting this
environment variable before starting the layout tests seems like
an acceptable workaround. Note that something similar is also
already done for the JSC tests.

* Scripts/webkitpy/port/base.py:
(Port.setup_environ_for_server):


Identifier: 230543@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268581 268f45cc-cd09-0410-ab3c-d52691b4dbfc