test262: test/language/expressions/conditional/in-branch-1.js https://bugs.webkit.org/show_bug.cgi?id=217879
Patch by Mark Cohen <m@mpc.sh> on 2020-10-19
Reviewed by Darin Adler.

JSTests:

I fixed this test262 failure - see the JSC ChangeLog for more details.
All I've done in this directory is remove the expected failure from
test262/expectations.yaml.

* test262/expectations.yaml:

Source/JavaScriptCore:

The test262 test in question checks that the parser respects the +In
parameter on the left-hand AssignmentExpression (between `?` and `:`)
in the ternary operator grammar. The relevant piece of the spec can be
found here (https://tc39.es/ecma262/#sec-conditional-operator). The
test checks this by embedding a ternary with left-hand
AssignmentExpression that contains the `in` keyword in the
initializing statement of a `for` loop, where `in` would normally be
disallowed. All this patch does is unconditionally allow the `in`
keyword inside the left-hand AssignmentExpression of a ternary.

This also fixes a variable typo in parseForStatement.

* parser/Parser.cpp:

Identifier: 230641@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268688 268f45cc-cd09-0410-ab3c-d52691b4dbfc