[GPU Process] RemoteAudioDestination::render() should not need to dispatch to the main thread to do IPC https://bugs.webkit.org/show_bug.cgi?id=217920
Reviewed by Alex Christensen.

Source/WebKit:

RemoteAudioDestination::render() should not need to dispatch to the main thread to do IPC as this is bad
for performance and IPC::Connection::sendWithAsyncReply() already does locking internally for thread-safety.
The only thing that was preventing us from calling sendWithAsyncReply() from the audio rendering thread was
a threading assertion in CompletionHandler that made sure that the completion handler was always called on
the thread the completion handler was constructed on. To avoid this issue, I added a paramater to the
CompletionHandler constructor allowing the caller to indicate it expects the completion handler to get called
on the main thread. I am now using this new flag in sendWithAsyncReply() and at its call site in
RemoteAudioDestination::render().

No new tests, no Web-facing behavior change.

* GPUProcess/media/RemoteAudioDestinationManager.cpp:
(WebKit::RemoteAudioDestination::render):
* Platform/IPC/Connection.h:
(IPC::Connection::sendWithAsyncReply):

Source/WTF:

Add an optional parameter to the CompletionHandler constructor allowing the caller to indicate it expects
the completion handler to get call on the main thread instead of the construction thread.

* wtf/CompletionHandler.h:
(WTF::CompletionHandler<Out):


Identifier: 230643@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268690 268f45cc-cd09-0410-ab3c-d52691b4dbfc