[GTK] Gamepad events do not occur in the flatpak shell. https://bugs.webkit.org/show_bug.cgi?id=212453
Reviewed by Michael Catanzaro.

If the support for gudev is disabled in libmanette, it falls back to monitoring the
files under /dev/input by itself without trying to use udev at all. This works fine
because the Flatpak SDK environment has access to all of /dev.

* buildstream/elements/sdk/libmanette.bst: Pass -Dgudev=disabled to Meson, and remove
the libgudev dependency from the element.


Identifier: 230672@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268721 268f45cc-cd09-0410-ab3c-d52691b4dbfc