Stop gap patch fix for regression in r267329. https://bugs.webkit.org/show_bug.cgi?id=218020 Source/WebCore:
<rdar://problem/69542459>

Reviewed by Darin Adler.

The refactoring in https://bugs.webkit.org/show_bug.cgi?id=216739 caused selections to
expand in both directions after a double-click and drag to expand on mac. This behavior
was not observed on iOS. This removes the error-prone calculation.
Note that this bug does not occur if only testing immediatly moving the drag point to the
end of the selection. Dragging across all the words as a real user would needs to be emulated.

Test: editing/selection/double-click-and-drag-over-anchor-to-select.html

* editing/VisibleSelection.cpp:
(WebCore::VisibleSelection::validate):

LayoutTests:

Reviewed by Darin Adler.

* editing/selection/double-click-and-drag-over-anchor-to-select-expected.txt: Added.
* editing/selection/double-click-and-drag-over-anchor-to-select.html: Added.
* resources/ui-helper.js:
(window.UIHelper.doubleClickAtMouseDown):
(window.UIHelper.mouseUp):
(window.UIHelper.dragMouseAcrossElement):
(window.UIHelper.doubleClickElementMouseDown):


Identifier: 230775@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268847 268f45cc-cd09-0410-ab3c-d52691b4dbfc